+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## [Unreleased]
 
+### Fixed
+- Crash when reading instruction file without parameter `hft.digestion.i_g_1992_ijk`.
+
 ### Changed
 - Single values in TOML file can be parsed like a single-element array.